Custom password masking character.
The default password masking character for the password input field was
changed to a spade symbol, which inturn is obtained from a custom font
in the fonts folder. Also the opacity of the sign up box was reduced a
little for better viewing the seven of spades card in the background.